% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon-wrappers.R
\name{AggregateSummaryAgeModels}
\alias{AggregateSummaryAgeModels}
\title{Title}
\usage{
AggregateSummaryAgeModels(top.dir.path)
}
\arguments{
\item{top.dir.path}{}
}
\value{

}
\description{
Title
}
\examples{
summary.age.mods <- AggregateSummaryAgeModels("inst/extdata/terr_14C_min10_dates-2020.03.04_15-19-42/") \%>\%
  tbl_df()

summary.age.mods \%>\%
  filter(DataName \%in\% sample(unique(DataName), 5)) \%>\%
  ggplot(aes(x = depth, y = mean, colour = DataName)) +
  geom_ribbon(aes(ymax = max, ymin = min, fill = DataName), colour = NA, alpha = 0.25) +
  geom_line() +
  geom_line(aes(y = median), linetype = 2)
}
