% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon-wrappers.R
\name{ConstructAgeModels}
\alias{ConstructAgeModels}
\title{Title}
\usage{
ConstructAgeModels(bacon.posterior, pars)
}
\arguments{
\item{bacon.posterior}{}

\item{pars}{}
}
\value{

}
\description{
Title
}
\examples{
posterior <- read.table("inst/extdata/test-subset/AE3/AE3_62.out", header = FALSE)
pars <- read.csv("inst/extdata/test-subset/AE3/bacon_pars.csv", stringsAsFactors = FALSE)


age.mods <- ConstructAgeModels(posterior, pars) \%>\%
  as_tibble()

age.mods \%>\%
  filter(iter \%in\% sample(unique(iter), 100)) \%>\%
  ggplot(aes(x = depth, y = age, group = iter)) +
  geom_line(alpha = 0.1) +
  expand_limits(x = 0, y = 0)
}
