% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon-wrappers.R
\name{CreateParameterFiles}
\alias{CreateParameterFiles}
\title{Add parameter files to a set of site directories}
\usage{
CreateParameterFiles(top.dir.path, pars.df)
}
\arguments{
\item{top.dir.path}{Path to top level directory for a set of site directories}

\item{pars.df}{The dataframe of Bacon parameters}
}
\value{

}
\description{
Add parameter files to a set of site directories
}
