% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon-wrappers.R
\name{GetAgeModAtDepths}
\alias{GetAgeModAtDepths}
\title{Title}
\usage{
GetAgeModAtDepths(age.mod, depth)
}
\arguments{
\item{age.mod}{}

\item{depth}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
